feat: set clean architecture basics and move some files
