+++ .env.example
@@ -1,0 +1,3 @@
+API_KEY=<PUT YOUR API KEY HERE>
+BASE_URL=https://api.yelp.com/v3/graphql
+IS_PRODUCTION=false
