+++ .fvmrc
@@ -1,0 +1,4 @@
+{
+  "flutter": "3.22.3",
+  "flavors": {}
+}
