+++ .gitignore
@@ -46,4 +46,10 @@ app.*.map.json
 /android/app/release
 
 # fvm
-.fvm/flutter_sdk
+
+# FVM Version Cache
+.fvm/
+
+
+# Others
+.env
