+++ lib/app/domain/entities/restaurant.dart
