+++ lib/app/domain/utils/constants.dart
@@ -1,0 +1,7 @@
+final class Constants {
+  const Constants._();
+
+  static const String apiKey = String.fromEnvironment('API_KEY');
+  static const String baseUrl = String.fromEnvironment('BASE_URL');
+  static const bool isProduction = bool.fromEnvironment('IS_PRODUCTION');
+}
