+++ lib/app/ui/features/app/app_widget.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+import '../home/home_page.dart';
+
+class AppWidget extends StatelessWidget {
+  const AppWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      title: 'Restaurant Tour',
+      home: HomePage(),
+    );
+  }
+}
