+++ lib/app/ui/features/home/home_page.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+
+import '../../../domain/entities/restaurant.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  Future<RestaurantQueryResult?> getRestaurants({int offset = 0}) async {
+    final headers = {
+      'Authorization': 'Bearer $_apiKey',
+      'Content-Type': 'application/graphql',
+    };
+
+    try {
+      final response = await http.post(
+        Uri.parse(_baseUrl),
+        headers: headers,
+        body: query(offset),
+      );
+
+      if (response.statusCode == 200) {
+        return RestaurantQueryResult.fromJson(
+          jsonDecode(response.body)['data']['search'],
+        );
+      } else {
+        print('Failed to load restaurants: ${response.statusCode}');
+        return null;
+      }
+    } catch (e) {
+      print('Error fetching restaurants: $e');
+      return null;
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            const Text('Restaurant Tour'),
+            ElevatedButton(
+              child: const Text('Fetch Restaurants'),
+              onPressed: () async {
+                try {
+                  final result = await getRestaurants();
+                  if (result != null) {
+                    print('Fetched ${result.restaurants!.length} restaurants');
+                  } else {
+                    print('No restaurants fetched');
+                  }
+                } catch (e) {
+                  print('Failed to fetch restaurants: $e');
+                }
+              },
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
